move sheet_to_html_string to sheets.views. delete sheets.utils.
