+++ sheets/views.py
@@ -5,6 +5,7 @@ import httplib2
 from StringIO import StringIO
 
 from django.template import RequestContext
+from django.template.loader import render_to_string
 from django.shortcuts import render_to_response, redirect
 from django.http import Http404
 
@@ -35,7 +36,6 @@ from sefaria.system.exceptions import InputError
 # noinspection PyUnresolvedReferences
 import sefaria.model.dependencies
 
-from sheets.utils import sheet_to_html_string
 from gauth.decorators import gauth_required
 
 
@@ -779,6 +779,43 @@ def make_sheet_from_text_api(request, ref, sources=None):
 	sources = sources.replace("_", " ").split("+") if sources else None
 	sheet = make_sheet_from_text(ref, sources=sources, uid=request.user.id, generatedBy=None, title=None)
 	return redirect("/sheets/%d" % sheet["id"])
+
+
+def sheet_to_html_string(sheet, request=None):
+	"""
+	Create the html string of sheet with sheet_id.
+	"""
+	sheet["sources"] = annotate_user_links(sheet["sources"])
+
+	try:
+		owner = User.objects.get(id=sheet["owner"])
+		author = owner.first_name + " " + owner.last_name
+	except User.DoesNotExist:
+		author = "Someone Mysterious"
+
+	sheet_group = (Group().load({"name": sheet["group"]})
+				   if "group" in sheet and sheet["group"] != "None" else None)
+
+	context = {
+		"sheetJSON": json.dumps(sheet),
+		"sheet": sheet,
+		"sheet_class": make_sheet_class_string(sheet),
+		"can_edit": False,
+		"can_add": False,
+		"title": sheet["title"],
+		"author": author,
+		"is_owner": False,
+		"is_public": sheet["status"] == "public",
+		"owner_groups": None,
+		"sheet_group":  sheet_group,
+		"like_count": len(sheet.get("likes", [])),
+		"viewer_is_liker": False,
+		"assignments_from_sheet": assignments_from_sheet(sheet['id']),
+	}
+
+	if request is not None:
+		return render_to_string('sheets.html', context, RequestContext(request))
+	return render_to_string('sheet_for_gdrive.html', context).encode('utf-8')
 
 
 @gauth_required(scope='https://www.googleapis.com/auth/drive.file')
